# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ARGOS",
    "description": "Agricultural Remote Gateway for Observation and Sensing",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/plants/{plant_id}": {
      "get": {
        "tags": [
          "plants"
        ],
        "summary": "Get Plant",
        "operationId": "get_plant_api_v1_plants__plant_id__get",
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantUnitRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "plants"
        ],
        "summary": "Update Plant",
        "operationId": "update_plant_api_v1_plants__plant_id__patch",
        "parameters": [
          {
            "name": "plant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Plant Id"
            }
          },
          {
            "name": "x-argos-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Argos-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlantUnitUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlantUnitRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/external/plants/{plant_code}": {
      "get": {
        "tags": [
          "external-plant-tracking"
        ],
        "summary": "Get External Plant",
        "operationId": "get_external_plant_api_v1_external_plants__plant_code__get",
        "security": [
          {
            "ExternalReadBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plant Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalPlantRead"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/api/v1/external/plants/{plant_code}/photos/{photo_id}/content": {
      "get": {
        "tags": [
          "external-plant-tracking"
        ],
        "summary": "Get External Photo Content",
        "operationId": "get_external_photo_content_api_v1_external_plants__plant_code__photos__photo_id__content_get",
        "security": [
          {
            "ExternalReadBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plant_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plant Code"
            }
          },
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExternalErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ExternalErrorDetail"
      },
      "ExternalErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ExternalErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ExternalErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlantUnitRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "public_code": {
            "type": "string",
            "title": "Public Code"
          },
          "species": {
            "type": "string",
            "title": "Species"
          },
          "species_label": {
            "type": "string",
            "title": "Species Label"
          },
          "variety": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety"
          },
          "rootstock": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rootstock"
          },
          "parcel_slug": {
            "type": "string",
            "title": "Parcel Slug"
          },
          "parcel_name": {
            "type": "string",
            "title": "Parcel Name"
          },
          "matrix_row": {
            "type": "integer",
            "title": "Matrix Row"
          },
          "matrix_column": {
            "type": "integer",
            "title": "Matrix Column"
          },
          "matrix_position_code": {
            "type": "string",
            "title": "Matrix Position Code"
          },
          "planted_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planted On"
          },
          "planted_on_precision": {
            "type": "string",
            "title": "Planted On Precision"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_label": {
            "type": "string",
            "title": "Status Label"
          },
          "irrigation_sector_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irrigation Sector Id"
          },
          "irrigation_line_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Irrigation Line Slug"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "public_code",
          "species",
          "species_label",
          "variety",
          "rootstock",
          "parcel_slug",
          "parcel_name",
          "matrix_row",
          "matrix_column",
          "matrix_position_code",
          "planted_on",
          "planted_on_precision",
          "status",
          "status_label",
          "irrigation_sector_id",
          "irrigation_line_slug",
          "latitude",
          "longitude",
          "notes",
          "created_at",
          "updated_at"
        ],
        "title": "PlantUnitRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "ExternalReadBearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "ExternalWriteBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}